Remove duplicate figure-1-person-token label
This label was present here and in github2gitlab.rst. In the latter it
actually makes sense, as it comes right before showing a figure
describing how to get a personal token in GitHub.

This generated a warning when building the documentation locally, and
surely generated warnings elsewhere too. The label was not referenced in
any place of the documentation though.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>